requirements.txt
flask
paho-mqtt
schedule
requests